Revert ObjectRootsFrameTests fix and change show/hideView() instead.
The ObjectRootsFrame is shown if a user right-clicks on a TreeItem in the
object browser. SwingUtilities.invokeLater() has been used earlier which causes
for VISIBLE events to be still in the event queue. What we really want here is
to wait for the VISIBLE event to be processed before we do any further actions.
This guarrantees that VISIBLE events are fired only once no matter what the
thread scheduling/event queue processing is.

This is a better approach, because we don't want VISIBLE events to fire
randomly as they might trigger certain actions, such as spawning a thread or
starting a timer.

EdtHelper and its test has been moved to the swing-components bundle.

Reviewed-by: omajid
Review-thread: http://icedtea.classpath.org/pipermail/thermostat/2012-July/002528.html