Lazy per-command-launch osgi bundle loading.
Reviewed-by: rkennke, omajid, sgehwolf, neugens
Review-thread: http://icedtea.classpath.org/pipermail/thermostat/2012-July/002364.html

It is wasteful and in some cases harmful to load every single Thermostat
bundle at beginning of program.  This patch refactors the launch process so
that a core set of 'bootstrap' bundles are initially loaded, and instead
the launcher is given a way to load dependency bundles before attempting
to execute a command.

Note: this patch breaks ActivatorTest from launcher bundle.  This is fixed
in a subsequent commit.  Please do not try to build/run this revision.