Add Eclipse feature and p2 repositories.
This patch adds a feature project which is required for installing things via
p2 (p2 is Eclipse's installer). Features can be thought of a collection of
plug-ins.

This feature is used to create one of the 3 p2 repositories:

1.) One repo is basically a repository of the core thermostat
    bundles with p2 metadata added. This is required for the
    Eclipse bits to properly resolve core thermostat bundles
    which it depends on.
2.) The second repo contains the Eclipse plug-in client.
3.) A composite repo which points to the above two sub-repos, plus
    eclipse.org's orbit repo (which we need for mongodb and lucene
    for now).

Here is how users can then try the Eclipse plug-in client:

1.) $ cd $THERMOSTAT_HOME
2.) $ mvn clean install
3.) $ cd eclipse
4.) $ mvn clean package
5.) "Help => Install new software", point it to
    $THERMOSTAT_HOME/eclipse/composite-repo/target/composite-repo
    and install the Eclipse prototype into their Eclipse.

Reviewed-by: neugens
Review-thread: http://icedtea.classpath.org/pipermail/thermostat/2012-August/002899.html