Split off agent and backend DAOs from storage
This has been a TODO for a while. Storage should not have to know
about the agent as much as possible. This patch extracts some bits
into AgentInfoDAO and BackendInfoDAO and reduces the interface
exposted by Storage.

Two separate DAOs are being used instead of one to keep the schema
as flat as possible.

Reviewed-by: vanaltj
Review-thread: http://icedtea.classpath.org/pipermail/thermostat/2012-September/003068.html