Expose all DAOs as services and use AgentInfoDAO as a service
Export all DAOs as services and update all users of AgentInfoDAO
(other than the startup bits that register the DAOs) to use that
instead of DAOFactory.

Try to handle services registering and unregistering appropriately.
This makes activators more complex which makes writing unit tests
harder. Introduce stub implementations of BundleContext (and
ServiceReference, ServiceRegistration and Filter) to address that.
These implementations are not complete, but they are sufficent for
the unit tests.

Reviewed-by: rkennke
Review-thread: http://icedtea.classpath.org/pipermail/thermostat/2012-September/003368.html