i18n strings in PingCommand and add some tests
Move strings from PingCommand into a resource bundle so the strings
can be localized. Add some unit tests for PingCommand (mostly just
checking the error strings).

Fix Translate so an instance of it can only be used with the enum
(normally a LocaleResources class) that it is initialized with.

Reviewed-by: jerboaa, neugens
Review-thread: http://icedtea.classpath.org/pipermail/thermostat/2012-October/003598.html