Thermostat Eclipse perspective, SWTComponent and metadata updates
This commit updates the dependencies in each Eclipse POM file. The new
charting capabilities rely on com.redhat.thermostat.client.ui. I have
changed the com.redhat.thermostat.eclipse plugin to specify a dependency
on thermostat-client-core, and the other eclipse projects to depend on
the main com.redhat.thermostat.eclipse plugin. This commit also adds the
new chart plugin as a friend of the main eclipse plugin.

The commit also contains a Thermostat perspective which collects all of
the Thermostat views in a manner similar to the Swing GUI, and a
constants interface that stores cross-plugin constants such as the
various view IDs.

I also added a SWTComponent subclass of UIComponent which contains a
createControl method. This is necessary since, unlike Swing, SWT
controls require a parent composite to be specified in their constructor.

Reviewed-by: jerboaa
Review-thread: http://icedtea.classpath.org/pipermail/thermostat/2012-September/003345.html