Remove OSGiContext interface.
In a previous review I got the impression that using OSGiContext is bad. Since
it was used by GuiClientCommand and ShellCommand only I thought we might as
well remove it. Similar to an earlier patch I've used FrameworkUtil to get at a
BundleContext and used static mocking for ActivatorTest in client-heapdumper.

This should reduce things which we special-case due to a command requiring a
bundle context.

Reviewed-by: neugens
Review-thread: http://icedtea.classpath.org/pipermail/thermostat/2012-October/003983.html