Clean up Eclipse client Maven dependencies
This commit cleans up a lot of duplicated Maven dependencies in the
Eclipse client's POM files. The Eclipse plugins are building using
Tycho, which resolves dependencies using the plugin's manifest. So the
only dependencies we need in the POM file are for core Thermostat
bundles. The required core Thermostat bundles are tracked in the client
feature, so it is much cleaner and more consistent to only specify a
dependency on the client feature in each Eclipse plugin POM.

Reviewed-by: jerboaa
Review-thread: http://icedtea.classpath.org/pipermail/thermostat/2012-November/004072.html