Fix log-handler log-level bound in LoggingUtils.
Earlier the log level of handlers was bound by their respective default config
(INFO for ConsoleHandler). This meant that log messages of a log level lower
than INFO got lost. With this patch the level of handlers is set to the same as
the root. An additional unit test has been added to assert this.

Reviewed-by: omajid
Review-thread: http://icedtea.classpath.org/pipermail/thermostat/2012-December/004397.html