Create VM Memory plugin
This commit moves the memory-stats-panel out of the client tree into a
separate plugin in the top-level directory of Thermostat. This commit
follows the previous plugin commits closely. The activator is split into
two where the client-core activator registers the VmInformationService
and the client-swing activator registers the ViewProvider. The
controller is moved to an internal package. The module and packages have
been renamed for consistency with the other plugins.

Reviewed-by: omajid, vanaltj, jerboaa
Review-thread: http://icedtea.classpath.org/pipermail/thermostat/2012-December/004388.html