Fix Eclipse build by lowering commons-codec version
The Eclipse build is currently broken due to the new dependency on
commons-codec in web-client. The version initially specified was 1.7,
but Eclipse only has access to 1.3 in its repos. This commit changes the
Thermostat build to use version 1.3. The version of commons-codec 1.3 on
Maven Central doesn't have OSGi metadata, so I have switched to the
Apache Servicemix package. Also some of the API used throughout
Thermostat isn't present in 1.3, but this new API is for convenience. I
have added some UTF-8 translating functions to StringUtils that mimic
the behaviour of these new methods.

Reviewed-by: vanaltj
Review-thread: http://icedtea.classpath.org/pipermail/thermostat/2012-December/004519.html