Fix broken commands and add tests
Fix some commands that are currently broken and add some tests to
ensure that the commands are available and can be invoked. The tests
are very basic. Most of the tests for VM don't even test actual data,
just that the output does not contain errors.

Reviewed-by: vanaltj
Review-thread: http://icedtea.classpath.org/pipermail/thermostat/2012-December/004585.html