Remove PROTOCOL config from db.properties.
As of now the storage command is inherently tied to mongo. As such it doesn't
make sense to make the protocol configurable (since there is only one valid
value anyway). Hence, I've removed this config option.

If we eventually end up needing this again we can re-introduce it. Right now, I
think it's wrong to ship a config value which isn't tunable. My gut feeling
tells me that we may never use it. In the sql case we will likely not start the
db itself via thermostat means. For the embedded case the DB may not need to be
started at all.

Reviewed-by: rkennke
Review-thread: http://icedtea.classpath.org/pipermail/thermostat/2012-December/004931.html