Add SSL (option) to storage command.
This is the first step towards adding TLS to mongo<-->agent|client|webservice
communication. It adds appropriate options to the mongod command if thermostat
is so configured. In order to test this better, I've refactored
MongoProcessRunner a bit. Also, DBStartupConfiguration does the parsing of the
db.properties file now. As such it's also easier to test. Tests have been added
for it too. Finally, I've renamed the c.r.t.agent.cli.db package to
c.r.t.agent.cli.impl.db to better reflect in the package name that it is an
internal package.

Note that you'll need to have a mongod available on your system which
understands --ssl* options in order to be able to fire mongod up with SSL
enabled. These options are only added if appropriate config is in place in
db.properties. Since it defaults to false, it should be OK to be pushed now
without breaking existing behaviour.

What comes next is adding support on agent/client/webservice side so that they
can talk SSL over the mongodb channel as well.

Reviewed-by: vanaltj, rkennke
Review-thread: http://icedtea.classpath.org/pipermail/thermostat/2012-December/004940.html
PR1243