Create Host CPU agent and common bundles
This first commit, in a series, extracts Host CPU data collection from
SystemBackend into a host-cpu-agent bundle. It also moves the CpuStatDAO
from common-core into a host-cpu-common bundle that registers the DAO
once Storage is available. This also removes the DAO from DAOFactory.

One side-effect is changing the visibility of HostLatestPojoListGetter
to public in order for this DAO implementation (and DAOs in the other
commits in this series) to access it.

Reviewed-by: omajid, jerboaa
Review-thread: http://icedtea.classpath.org/pipermail/thermostat/2013-January/004993.html