Merge bundles/launcher plug-ins.
Here is a patch which merges classes which used to be in the bundles plug-in
into the launcher plug-in. I don't see why this splitting is useful any longer.
Now we have the "main" jar which fires up the framework and uses the launcher
plug-in to do the osgi-based launching. This is a first step to clean up
dependencies amongst bundles which should eventually lead to cleaner dep-trees
and common-core not depend on storage-core any longer, having this c.r.t.test
package removed from common-core etc.

ThermostatTest was also largely commented out. The only non-comment part was
the setup routine. A few basic tests have been reinstated.

Reviewed-by: vanaltj
Review-thread: http://icedtea.classpath.org/pipermail/thermostat/2013-January/005093.html