Introduce Service and ExtensionPoint annotations
Thermostat currently has a large number of interfaces that are meant to
be used as services (where a client obtains an instance of the interface
from OSGi) or as extension points (where a client registers an instance
of the interface as an OSGi service and thermostat makes use of it).
There is no clear indication of which interfaces are what.

Clarify this by introducing two annotations @Service and @ExtensionPoint
that should be used to document the purpose of the interface. These
annotations do not imply any behaviour and are for documentation
purposes only.

Also introduce an annotation processor that writes "plugin-docs.xml"
files containing some basic documentation about classes marked with
@Service or @ExtensionPoint. Nothing explicit needs to be done to
invoke the annotation processor, simply including a dependency on
thermostat-annotations (needed to use @Service/@ExtensionPoint anyway)
is enough. These xml files are not included in the jars.

Reviewed-by: ebaron, jerboaa, vanaltj
Review-thread: http://icedtea.classpath.org/pipermail/thermostat/2013-January/004983.html
PR1255