Change BackendInformation to implement Ordered
This simple commit follows up on a TODO I previously placed to sort
Backends using the OrderedComparator, which is done for
InformationServices. This was impossible before due to a cyclic
dependency between common-core and storage-core, but this has since been
resolved.

Reviewed-by: omajid
Review-thread: http://icedtea.classpath.org/pipermail/thermostat/2013-February/005643.html