Use bundle version in SystemBackend
I have noticed that SystemBackend is reporting 0.5.0 as its version
number. It turns out this is still using a hard-coded version string,
while the other Backends use the version of their containing bundle.
This commit changes SystemBackend to use the bundle version and includes
a test for this as well.

Reviewed-by: omajid
Review-thread: http://icedtea.classpath.org/pipermail/thermostat/2013-February/005850.html