Agent Proxy client
This patch implements the client side of the Agent Proxy that integrates
into Thermostat. The Agent's activator now registers an AgentProxyFilter
with the VmBlacklist service from the previous patch. This Filter causes
Thermostat to not monitor the Agent Proxy process for reasons outlined
in the previous email.

The bulk of the functionality is in the AgentProxyClient class, which is
responsible for creating the Agent Proxy process and communicating with
it using RMI. This class performs a sort of handshake with the Agent
Proxy server, using the AgentProxyListener interface, in order to
determine when the server is up and exported to the RMI registry. Once
the server is available, the client retrieves its AgentProxyLogin
interface and attempts to log in. If the login is successful, an
AgentProxyControl interface is returned that allows the client to
perform privileged operations on the server. These operations are:
attach, isAttached, getConnectorAddress, and detach. The MXBeanConnector
class now delegates much of its previous functionality to the
AgentProxyClient class.

The MXBeanConnectionPoolImpl class now manages a new RMIRegistry class,
which encapsulates functionality like creating the registry, exporting
and unexporting objects. Typically these are done with static methods,
so this RMIRegistry class makes testing much easier, in addition to
managing the lifecycle of the actual RMI registry. The RMI registry is
created with a custom ServerSocket factory that binds to the loopback
address, making sure only local connections are permitted.

Reviewed-by: jerboaa
Review-thread: http://icedtea.classpath.org/pipermail/thermostat/2013-November/008704.html
PR1460