Fix agent proxy startup issue
This patch fixes a bug where the agent proxy cannot start due to the
recent addition of the CommonPaths OSGi service. Since the agent proxy
is not running under OSGi, CommonPaths is not set in
NativeLibraryResolver as it normally is in Thermostat itself.

Reviewed-by: omajid
Review-thread: http://icedtea.classpath.org/pipermail/thermostat/2013-November/008844.html