Command Channel as a separate process
This patch moves the command channel server to a separate process. The
motivation behind this is that the command channel server listens over
the network for incoming commands. If the Thermostat agent is being run
as root, for example to monitor all users' VMs, then a
wide-network-facing open port could present an attack vector.

Reviewed-by: omajid, jerboaa, aazores
Review-thread: http://icedtea.classpath.org/pipermail/thermostat/2015-August/015411.html