Packaging fixes for command channel
This patch modifies our spec file to install the
thermostat-command-channel script such that it is owned by the
thermostat user and group. This patch also adds back netty as an
explicit dependency in agent/command, and updates the Fedora patch to
reflect this change. Netty is listed as a dependency in the parent POM,
which is why Thermostat still builds.

Reviewed-by: jerboaa
Review-thread: http://icedtea.classpath.org/pipermail/thermostat/2015-September/016083.html